# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API is used to send notification.",
        "title": "Notificcation API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth": {
            "post": {
                "description": "Apply a authrization token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Apply a authrization token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Use the zilliz email",
                        "name": "user",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "application name",
                        "name": "app",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/email": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Send a message to a specify email address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Send"
                ],
                "summary": "Send message by email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "email address",
                        "name": "receiver",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "email subject",
                        "name": "subject",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "email message",
                        "name": "message",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "email content format, text or html, default text",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/feishu": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Send a message to feishu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Send"
                ],
                "summary": "Send message by feishu",
                "parameters": [
                    {
                        "type": "string",
                        "description": "email address",
                        "name": "receiver",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "email message",
                        "name": "message",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/list": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "List all message template",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Template"
                ],
                "summary": "List message template",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/refresh": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Refresh Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh Token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/sms": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Send a message to a phonenumber",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Send"
                ],
                "summary": "Send message by sms",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Message Template Id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Message Template Name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Message Params",
                        "name": "params",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Receiver phone number, area code required",
                        "name": "receiver",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}